+++ one-connected-code-biconnected-time-complexity/biconnected.hpp
@@ -10,18 +10,18 @@ class FaceTriangulation;
 class biconnected
 {
 public:
-    vector<vector<int>> faces;
-    unordered_multiset<pair<int, int>, PairHash> present;
-    vector<vector<pair<int, int>>> allTriangulations;
+    vector<vector<long long>> faces;
+    unordered_multiset<pair<long long, long long>, PairHash> present;
+    vector<vector<pair<long long, long long>>> allTriangulations;
     vector<FaceTriangulation*> faceTriangulations;
-    int totalTriangulations;
-    int totalChecks;
-    int successfulChecks;
-    int invalidTraversals;
-    biconnected(vector<vector<int>> &faces)
+    long long totalTriangulations;
+    long long totalChecks;
+    long long successfulChecks;
+    long long invalidTraversals;
+    biconnected(vector<vector<long long>> &faces)
     {
         this->faces = faces;
-        present = unordered_multiset<pair<int, int>, PairHash>();
+        present = unordered_multiset<pair<long long, long long>, PairHash>();
         initiatePresent();
         faceTriangulations = vector<FaceTriangulation*>(faces.size());
         totalChecks = 0;
@@ -33,7 +33,7 @@ public:
     {
         for (auto face: faces)
         {
-            for(int i = 0 ; i < face.size() - 1; i++)
+            for(long long i = 0 ; i < face.size() - 1; i++)
             {
                 present.insert(make_pair(min(face[i], face[i+1]), max(face[i], face[i+1])));
             }
@@ -41,7 +41,7 @@ public:
         }
     }
     void getAllTriangulations();
-    void output(int serial);
+    void output(long long serial);
     void addTriangulation();
     void sortTriangulations()
     {
@@ -79,7 +79,7 @@ inline void biconnected::getAllTriangulations()
     // printAllTriangulations();
 }
 
-inline void biconnected::output(int serial)
+inline void biconnected::output(long long serial)
 {
     if (serial == faces.size() - 1)
     {
@@ -97,10 +97,10 @@ inline void biconnected::output(int serial)
 
 // inline void biconnected::addTriangulation()
 // {
-//     vector<pair<int, int>> currentTriangulations;
+//     vector<pair<long long, long long>> currentTriangulations;
 //     for (auto a : faceTriangulations)
 //     {
-//         vector<pair<int, int>> currentTriangulation;
+//         vector<pair<long long, long long>> currentTriangulation;
 //         for (auto &chord : a->chords)
 //         {
 //             currentTriangulation.push_back(a->getPair(chord));
